moved dbhandler files to the pir directory
